Refactor sound code to it's own folder, delete Pi::BoinkNoise.
