+++ src/LuaEngine.cpp
@@ -21,12 +21,12 @@
 #include "Player.h"
 #include "Random.h"
 #include "SectorView.h"
-#include "Sound.h"
-#include "SoundMusic.h"
 #include "WorldView.h"
 #include "buildopts.h"
 #include "graphics/Graphics.h"
 #include "scenegraph/Model.h"
+#include "sound/Sound.h"
+#include "sound/SoundMusic.h"
 #include "ui/Context.h"
 #include "utils.h"
 /*
