+++ src/Player.cpp
@@ -13,11 +13,11 @@
 #include "Pi.h"
 #include "SectorView.h"
 #include "ShipCpanel.h"
-#include "Sound.h"
 #include "SpaceStation.h"
 #include "StringF.h"
 #include "SystemView.h" // for the transfer planner
 #include "WorldView.h"
+#include "sound/Sound.h"
 
 //Some player specific sounds
 static Sound::Event s_soundUndercarriage;
