+++ src/sound/AmbientSounds.cpp
