+++ src/sound/SoundMusic.cpp
