+++ src/sound/SoundMusic.h
